Disable anchor link in same page
The site has some kind of weird behaviour (some javascript?) with
anchors. When referring an anchor, it gets stripped from the url.

But with the anchor is in the same page, the client gets redirected
to the previous page (???).

I do not want to block the change so I disable the link for the time
being.