Remove yamllint gem.
In other projects, we use the Python yamllint tool. When this gem is
installed, it can come first in the path, and conflict causing these
tests to fail.

This was originally added when this was ported to sinatra in 6863ad1,
but doesn't seem to be used at all. It therefore makes sense to remove.
If we want to do yaml linting, we should use the same tool as in other
projects to avoid confusion.