GenericContact form for contact-us and signup
We want to not repeat ourselves but implement a generic
GenericContact form model that we specialise for each case.

GenericContact would render the message and generate the ticket
object for zendesk based on the existing fields.

We refactor Contact and Signup accordingly.

Signup overrides the `rendered_message()` method to add the
invites.

Signed-off-by: Hector Rivas Gandara <keymon@gmail.com>