Add expectations for ticket created in ZenDesk
We want to verify that we do the corresponding request to ZenDesk
to create the tickets. For that we leverage WebMock and add several
expectations in the contact_us and signup form.

We remove the sinatra app we were using to implement the zendesk
fake API, as there is no need for it if we use the provided
zendesk mocking function.

Signed-off-by: Hector Rivas Gandara <keymon@gmail.com>