Implement tests for all support subforms
We add the tests for all the support subforms. We try to
use table tests to avoid repetition in the tests: We fill each
form in a `before(:each)` and wipe each field as required.

We were not able to "unselect" an option for the first form,
but instead we fill the form again.

Signed-off-by: Hector Rivas Gandara <keymon@gmail.com>