Increase the form message size to 50k
We want to allow the users to put more data than 2k in the message
field. We increase the max message length to 50000, as the
maximum comment size in ZenDesk is 64k[1] and we need space for
other info.

In a follow up story we should test this and add some feedback to
the user about the maximum size.

[1] https://support.zendesk.com/hc/en-us/articles/360000567348-Is-there-a-character-limit-on-ticket-comments-